refactor: added missing elements to drafts page
