+++ pages/index.tsx
@@ -3,7 +3,6 @@ import useTokenGated from "lib/useTokenGated";
 import Layout from "components/Layout";
 import Navbuttons from 'components/Navbuttons'
 
-
 import {
   Box,
   Button,
@@ -20,6 +19,7 @@ import { FiPlus } from "react-icons/fi";
 import Link from "next/link";
 import { useProposals } from "lib/useProposals";
 import { Proposal } from "@prisma/client";
+import ProposalButton from "components/NewProposalButton";
 
 const ProposalRow = (props: Proposal) => (
   <Box p="20px">
@@ -38,14 +38,9 @@ const Page: NextPage = () => {
     <Layout>
       <VStack align="stretch" spacing={5}>
         <Flex>
-          {/* Sending the page prop to the Navbuttons.tsx to get the active button  */}
           <Navbuttons page="All"/>
           <Spacer />
-          <Link href="/new" passHref>
-            <Button colorScheme="blue" leftIcon={<Icon as={FiPlus} />}>
-              New Proposal
-            </Button>
-          </Link>
+          <ProposalButton />
         </Flex>
 
         <VStack
